+++ art/__init__.py
@@ -1,2 +1,3 @@
 """Art modules."""
 from .art import *
+__version__ = ART_VERSION